requirements.txt
tqdm==4.65.0
yara_python_dex==1.0.1